+++ spotinst_sdk2/clients/managed_instance/__init__.py
@@ -289,3 +289,69 @@ class ManagedInstanceAwsClient(Client):
 
         return formatted_response['response']
     
+    def import_multiple_instances_from_aws(self, migrations_list: list):
+        """
+        Import Multiple Instances From AWS
+
+        # Arguments
+        migrations_list (List): ManagedInstanceMigration List
+
+        # Returns
+        (Object): ManagedInstanceMigrations API response
+        """
+        req = aws_managed_instance.ManagedInstanceBulkMigrationRequest(migrations_list)
+
+        post_url = self.__base_mi_url + "/" + "migration" + "/" + "bulk"
+
+        req_exclude_missing = self.exclude_missing(json.loads(req.toJSON()))
+
+        formatted_mi_dict = self.convert_json(
+            req_exclude_missing, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_mi_dict)
+
+        mi_response = self.send_post(
+            body=body_json,
+            url=post_url,
+            entity_name=self.ENTITY_NAME)
+
+        formatted_response = self.convert_json(
+            mi_response, self.camel_to_underscore)
+
+        ret_val = formatted_response["response"]["items"]
+
+        return ret_val
+    
+    def get_multiple_instances_migration_status(self, migration_ids: list):
+        """
+        Import Multiple Instances From AWS
+
+        # Arguments
+        migration_ids (List): List of ManagedInstanceMigrationIds 
+
+        # Returns
+        (Object): ManagedInstanceMigrations API response
+        """
+        req = aws_managed_instance.ManagedInstanceBulkMigrationStatusRequest(migration_ids)
+
+        post_url = self.__base_mi_url + "/" + "migration" + "/" + "search"
+
+        req_exclude_missing = self.exclude_missing(json.loads(req.toJSON()))
+
+        formatted_mi_dict = self.convert_json(
+            req_exclude_missing, self.underscore_to_camel)
+
+        body_json = json.dumps(formatted_mi_dict)
+
+        mi_response = self.send_post(
+            body=body_json,
+            url=post_url,
+            entity_name=self.ENTITY_NAME)
+
+        formatted_response = self.convert_json(
+            mi_response, self.camel_to_underscore)
+
+        ret_val = formatted_response["response"]["items"]
+
+        return ret_val
+